svstl/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
inflection==0.5.1
mysqlclient==2.2.0
outcome==1.2.0
packaging==23.1
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
selenium==4.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.4
webdriver-manager==4.0.0
wsproto==1.2.0
